+++ pkg/providers/amifamily/ami_test.go
