+++ pkg/providers/amifamily/bootstrap/bootstrap.go
