+++ pkg/providers/amifamily/bootstrap/bottlerocket.go
