+++ pkg/providers/amifamily/bootstrap/bottlerocketsettings.go
