+++ pkg/providers/amifamily/bootstrap/custom.go
