+++ pkg/providers/amifamily/bootstrap/eksbootstrap.go
