+++ pkg/providers/amifamily/custom.go
@@ -21,7 +21,7 @@ import (
 	"github.com/aws/karpenter-core/pkg/cloudprovider"
 
 	"github.com/aws/karpenter/pkg/apis/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudprovider/amifamily/bootstrap"
+	"github.com/aws/karpenter/pkg/providers/amifamily/bootstrap"
 )
 
 type Custom struct {
